refactor(test): consolidate cloudserver tests into cloudserver_resource_test.go
